[refactor] custom hook 만들고 정리하기
