+++ src/utils/api.js
@@ -1,7 +1,7 @@
 import { handleSessionExpired } from "./authentication";
 export async function api(endpoint, { body, ...customConfig } = {}) {
-  const SERVER_URL = "https://piro-recruiting.top/api";
-    // const SERVER_URL = "http://localhost:3000/api"; 
+  // const SERVER_URL = "https://piro-recruiting.top/api";
+    const SERVER_URL = "http://localhost:3000/api"; 
   const headers = { "Content-Type": "application/json" };
 
   const token = sessionStorage.getItem("token");
@@ -56,8 +56,8 @@ api.put = function (endpoint, body, customConfig = {}) {
 };
 
 api.sse = function (endpoint) {
-  const SERVER_URL = "https://piro-recruiting.top/api";
-    // const SERVER_URL = "http://localhost:3000/api"; 
+  // const SERVER_URL = "https://piro-recruiting.top/api";
+    const SERVER_URL = "http://localhost:3000/api"; 
   const token = sessionStorage.getItem("token");
   const url = `${SERVER_URL}${endpoint}?token=${token}`;
